Minor refactor, clean up unused imports
